==> 03_analysis_and_visualization.ipynb <==
pip install --upgrade seaborn matplotlib pandas

import pandas as pd
import os
import seaborn as sns
import matplotlib.pyplot as plt

from google.colab import drive
drive.mount('/content/drive')


# ----------------------------------------------------------------------
# 1. CONFIGURAÇÃO DE CAMINHOS E CARREGAMENTO
# ----------------------------------------------------------------------

# ⚠️ AJUSTE ESTE CAMINHO para o local da sua pasta no Drive!
# Deve ser o mesmo caminho usado no script de unificação (01).
BASE_PATH = '/content/drive/MyDrive/Coursera_google_analytics/'

# Caminho do arquivo unificado gerado no script anterior
UNIFIED_FILE_PATH = os.path.join(BASE_PATH, 'cyclistic_dados_limpos_analise.csv')

print(f"Carregando arquivo unificado bruto: '{UNIFIED_FILE_PATH}'")

# Carregar o arquivo unificado
# O low_memory=False ajuda a garantir que o Pandas identifique corretamente os tipos de dados
df = pd.read_csv(UNIFIED_FILE_PATH, low_memory=False)

print(f"✅ DataFrame 'df' carregado com sucesso. Total de linhas: {len(df)}")
print("Iniciando a limpeza de dados e criação de campos analíticos...")

# Calcula a média de minutos de todas as viagens
df["ride_length(min)"].mean(axis=0)

#Calcula a quantidade de horas por tipo de membro
ride_length_sum = df.groupby('member_casual')['ride_length(hours)'].sum()
print(ride_length_sum)

#Contagem total de viagensp or membro casual ou anual
day_week_count = df.groupby(['member_casual'])['member_casual'].count()
day_week_count

#Calcula a media de viagens em minutos por tipo de membro
df.groupby('member_casual')['ride_length(min)'].mean()

#Calcula o dia da semana que mais tem viagens
df["day_of_week_name"].mode()

#Calcula o dia da semana que mais tem viagens por membro
df.groupby('member_casual')['day_of_week_name'].apply(lambda x: x.mode())

# 1. Filtra apenas os clientes 'casual'
df_casual = df[df['member_casual'] == 'member']

# 2. Agrupa por estação e conta o número de viagens (ride_id)
# 3. Ordena a contagem de forma decrescente
top_casual_stations = df_casual.groupby('start_station_name')['ride_id'] \
                                .count() \
                                .sort_values(ascending=False)

# Exibe as 10 primeiras
print(top_casual_stations.head(10).to_frame(name='Total Viagens Casuais'))
print("-" * 50)
